+++ src/main/java/com/example/postapp/service/PostServiceImpl.java
@@ -74,8 +74,8 @@ public class PostServiceImpl implements PostService {
     // 게시글 삭제 http://localhost:8080/api/v1/posts/list/2
     @Transactional(readOnly = false)
     @Override
-    public void removePost(int id) {
-        postrepository.deleteById(id);
+    public void removePost(Long id) {
+        postrepository.deleteById(id.intValue());
     }
 
 }
